refactor: move individual book route to seperate file
